requirements.txt
pyyaml==5.3
pyinstaller==3.6
